+++ core/src/raw/oio/write/mod.rs
@@ -15,31 +15,9 @@
 // specific language governing permissions and limitations
 // under the License.
 
-//! `oio` provides OpenDAL's raw traits and types that opendal returns as
-//! output.
-//!
-//! Those types should only be used internally and we don't want users to
-//! depend on them. So we should also implement trait like `AsyncRead` for
-//! our `output` traits.
-
-mod read;
-pub use read::*;
-
-mod write;
-pub use write::*;
-
-mod append;
-pub use append::*;
-
-mod stream;
-pub use stream::*;
-
-mod page;
-pub use page::*;
-
-mod cursor;
-pub use cursor::Cursor;
-pub use cursor::VectorCursor;
-
-mod entry;
-pub use entry::Entry;
+mod api;
+pub use api::BlockingWrite;
+pub use api::BlockingWriter;
+pub use api::Write;
+pub use api::WriteOperation;
+pub use api::Writer;
